+++ src/propagation/check-for-decay.ts
